+++ samples/assistant/powershell/CreateAssistant/run.ps1
@@ -9,7 +9,9 @@ $instructions += "\nAsk for clarification if a user request is ambiguous."
 
 $create_request = @{
     "id" = $assistantId
-    "instructions" = $instructions
+    "instructions" = $instructions,
+    "chatStorageConnectionSetting" = "AzureWebJobsStorage",
+    "collectionName" = "SampleChatState"
 }
 
 Push-OutputBinding -Name Requests -Value (ConvertTo-Json $create_request)